efbp: Rework to support organic network halt
The network Pump now learns of Producers and Consumers in the network
and listens for shutdown events from these components. Once all
registered Producers have been shut down, a ShutdownCommand is
propagated through the network to the Consumers. Once all Consumers have
registered the shutdown event, the HaltCommand is propagated to the
Bootstrap component and execution comes to an end.