schedule: Shutdown Timer thread
Implementing example applications exposed an issue not seen in the test
suite where a network would halt but the application would remain alive,
i.e., control was not returned to the parent shell after execution.

The cause was the Timer instance starts a non-daemon thread which kept
the JVM alive, requiring some adjustment to the Trigger mechanism in the
DefaultPump to make its action more generic. Requiring a Runnable
instance rather than a Command allows the Trigger instance to perform
arbitrary actions on instances whose classes both do and do not
implement Node.